requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai-whisper==20231117
edge-tts==6.1.9
pdfminer.six==20221105
spacy==3.7.2
requests==2.31.0
python-multipart==0.0.6